Change FOL representation ; translate everything about data types
This commit is a big remake of the representation of first order
logic. It is now tailored for our uses, with min and CF as built-ins,
as well as pointers, projections and pointer application. Futhermore,
if you create formulae in cnf, even with the use of implications, you
can have them written in that format.

Further work includes a sound translation from Clauses Var -> Clauses
String (or SDoc), where we can shorten the variable names but in the
same time ensure that there is no collision.

There is support of having comments in the outputed TPTP, and it would
be nice to have the ghc core definition of a function together with
its translation to logic.