Update how contracts are translated
Translation is now done in one function trContract, which has a
negative or positive Variance argument.

We also telescope the current contract, so if we want to translate this contract:

    [[ e \notin (x : C1) -> (y : C2) -> CR ]]

It is not translated as this

    \exists x y . [[ x \in C1 ]] & [[ y \in C2 ]] & [[ e x y \notin CR ]]

i.e. the difference is that C1 and C2 are bound under the same exists.

This allows new possibilities to split this under several clauses,
since they are conjuncts. New skolem variables a_x and a_y will be
introduced, and we get clauses something like this:

   fof(_,axiom, [[ a_x \in C1 ]]).
   fof(_,axiom, [[ a_y \in C2 ]]).
   fof(_,axiom, [[ e a_x a_y \notin CR ]]).

We also need to be careful now, since a (:->) can have the same
binding variable to the right. This example is from PredLog.hs:

    map_retains_invariant =
        map ::: retain invariant --> retain (all invariant)

    -- | What it means to retain a predicate
    retain :: (a -> Bool) -> Contract (a -> a)
    retain p = Pred p :-> \x -> Pred (\ r -> p x && p r)

So the big contract for map is really

    map ::: (Pred invariant :-> \x -> Pred (\ r -> invariant x && invariant r))
        --> (Pred (all invariant) :-> \x -> Pred (\ r -> (all invariant ... )))

As the inliner currently works, the exactly the same `x` is used in
both occurences of :-> \x -> ... above. This is very bad now, since
the latter will overshadow the former with the telescoping as we have.
The solution is to let Contracts.Collect refresh the unique in all
lambda-bound variables after :->. I first tried to make the inliner to
do this, but it annoyingly made it live in a UniqSM monad.