Change the internal representation for contracts
In particular, we now support assumptions, (:=>). They are almost as
Using, but they stay when Using the contract. So we write

    eq_refl x = x ::: CF :=> x == x ::: CF :&: Pred id

Now, when `Using` eq_refl, the assumption that (x ::: CF) is copied as
well, which wouldn't happen if it instead was introduced with `Using`.

A peculiar bug is that the following contract does not hold when using
min translation:

    unused_false_assumption = bad ::: CF :=> False ::: Pred id

The reason is that while we say that BAD is CF, it is only the case
if we have min(BAD), but the consequent of the implication is
countersatisfied without ever making BAD min.