Readd splitting goals with the extended statement language
A very important thing was noticed. If you split the goal for a
function f, and that function is mentioned either in the predicate or
in an assumed contract, the entire definition still needs to be added.

The contracts collector has been simplified, the dependency analysis
for statements has been factored out.