Substitute fixed point induction functions over global Ids
If a function is imported from another module, its Id is global, and
the normal substitution functions does not touch them. The new
functions substGblId(s) are used from halo when fixating functions for
fpi or substituting in statements.

This bug was around in the last commits, so the examples in Length.hs
are again satisfiable.