Removing some nonsense from previous experiments and fixing some bugs.
The most important modification is that I have droped
some min requirements from function definitions. Mins
are still there to propagate strictness, but equalities
hold unconditionally. I believe this is important for
completeness (which I still have not proved but seems
to require a translation like this). In particular now
if we do evaluate g(x) and f(x) = g(x) then we will also
have f(x) in min. Hence, if f(x) eventually is not in min,
this means that neither it, nor something that is equal to
it has been called, hence we can equate ALL of them to bottom.
That's the point.