Committing a light version of the testsuite that does not contain any CF contracts, in order to experiment completeness and finite models without the CF-related complications. The examples are taken from new-testsuite/
There is already an interesting finding: the append contracts
hold (and 'minrec' is able to say UNSAT), but they appear to
be non-provable (SAT) in master branch. I don't really know
what to say, but there must be something incomplete in master.

NB: 'minrec' vs 'min' does not play a role here -- minrec only
has to do with CF and there are no CF contracts in this testsuite.