Implemented a flag for adding or not adding a min guard on arrow contracts, in order to run some experiments.
Also implemented a slight variation of the positive translation
namely:

  (min(e) /\ e = UNR )
  \/
  (min(p(e)) /\ (p(e) = UNR \/ p(e) = True))

which seems better from a finite model perspective (i.e. in combination
with absent min() guards from arrow contracts) since it does not ALWAYS
assert that p(f(x)) is in min! This allows the model finder to equate
finitely many f(x) to non-bottom values which is more likely to cause
min(p(f(x))) to introduce a finite number of elements in the model
(unless of course 'p' is one of these weird predicates that insert more
and more elements, see PInf.hs in testsuite/ -- that situation is hopeless
anyway)